bitset: clear stale bits in reset/resize
There's a gap in `mbitset_resize()` where the first limb of a grown
bitset is not initialized to zero.  This is most easily exposed when
using `mbitset_reset()` followed by `mbitset_resize()`:

```c
  M_LET(set, bitset_t) {
    bitset_init(set);
    bitset_resize(set, 128);
    bitset_set_at(set, 0, true);
    bitset_reset(set);
    bitset_resize(set, 128);
    assert(bitset_get(set, 0) == false);
  }
```

`mbitset_reset()` assumes the first limb has been zeroed either in
`realloc()` or by a previous `mbitset_resize()` call.  However,
`mbitset_reset()` reduces the size, but does not zero the first limb as
would be done by `mbitset_resize(set, 0)`.

In this commit, I have added the zeroing of the first limb to
`mbitset_reset()`.

NOTE: This fix does not resolve the issue if `realloc()` returns dirty
memory.  It could be done by calling `memset()` following the
allocation, but as the existing code avoids `memset()` I wasn't sure if
this was being intentionally avoided.